Move typings test file to root so its not run by mocha
